backend/ml/requirements.txt
accelerate==0.31.0
aiokafka==0.10.0
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
diffusers==0.28.2
filelock==3.14.0
fsspec==2024.6.0
grpcio==1.64.1
grpcio-tools==1.64.1
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
MarkupSafe==2.1.5
minio==7.2.7
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pillow==10.3.0
protobuf==5.27.0
psutil==5.9.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.1
zipp==3.19.2
translate
peft
rembg
matplotlib
pillow
pandas